[Add][Core] Connection type to hosts.yml
We are in a chicken and egg standoff over tests for this addition:

Issue 29 allows the test to be written.
Yet tests break in CICD over a ansible-inventory requirement.

This issue removes that requirement for docker use cases.
But any test will break in the CICD environment....

Note to future self:
Start every project with a working CICD environment.

Signed-off-by: Begley Brothers Ltd <begleybrothers@gmail.com>