Move 'bgp.address_families' default value into SR OS device data
bgp.address_families is an experimental attribute that is used to
implement IPv4-only network on top of SRv6 on SR OS. As the attribute
is not used in any other configuration template, the default value
has been moved into SR OS device data.

You can still specify global value for this attribute (and it will
be then copied into all node data), but the node data for all other
devices will not be polluted with unused default values.