Update routed VLAN subinterface name in the VLAN neighbor list (fixes #300)
VLAN neighbor list is built before the subinterfaces are renamed. That
works nicely for VLAN/BVI interfaces because they are created
on-the-fly, but not for routed VLAN subinterface that are renamed later
in the process.

This change calls 'update_vlan_neighbor_list' after a routed interface
has been renamed, passing fake parent interface with old interface name.

Neighbor list on that interface is empty, so there's no change in the
list of neighbors. The rest of the 'update_vlan_neighbor_list' then
deals with renamed interface.