Bug fix: reorder steps in early module initialization code
We need the ability to change system defaults of other modules in the
module initialization code (example: EVPN module has to change the
default setting of 'vxlan.flooding'). To get there, we had to rearrange
the sequence of operation in early module initialization.

All the tests are still OK, so I'm hoping this change did not break
anything and keeping my fingers crossed ;)