Reorganize ebgp.utils Cisco IOS template
Original template focused on individual functions and would generate
tons of 'address-family X' configuration commands. Ansible might not
like that when trying to figure out the minimum set of changes to
send to the device.

Reorganized template follows the Cisco IOS configuration hierarchy
resulting in minimal amount of context switching.