Fix routed native VLAN handling (fixes #456 #457 #461)
The 'routed native VLAN' scenario triggers cleanup similar to other 'routed
VLAN' cases early in the VLAN processing, leaving too little information for the
final processing phase where we need to decide whether to make an interface a
switchport or a routed port. The early processing now sets _vlan_native and
_vlan_mode internal interface attributes to make our life simpler.

The final VLAN physical interface processing removes routed native VLAN from
VLAN trunk (the proof of why we can do that is left as an exercise for the
reader) and triggers an error if the device does not support that
(vlan.native_routed feature). If we're dealing with pure L3 interface, removing
of native VLAN from VLAN trunk empties the trunk, and the subsequent checks and
configuration templates succeed.