Bug fix: use static address on Cumulus Linux mgmt interface
Background: 'ifreload -a' wouldn't work in Cumulus Linux containers due
to weird error messages reports by DHCP client. The workaround was to
create ever-more-complex interface configurations and use 'ifreload
--allow' hoping for the best.

The root cause was AppArmor DHCP profile. Removing that solved the
problem, but it's a bit of a hammer-of-Thor approach. Now that we have
static containerlab management addresses, we can change the management
IP in /etc/network/interfaces and get rid of DHCP problems for good.