vMX: Support for VLAN
Tested with VLAN integration test cases, including Router-on-a-Stick.

Additionally, changed the default interface behaviour with the help of device quirk modules.
Removed specific entities for "unit 0" on the initial interface configuration,
appending .0 to the "plain" interfaces on the quirk module.

The reason for that change is the ability to easily handle different subunits.