Add include-at-parent-level functionality, use it for VRF and VLAN data types
Included files could not add data above their inclusion point (which makes
sense ;), which meant that we had to default VRF and VLAN data types in the
global attributes.yml file.

This patch adds the capability to add included data within the _top key
to the parent data structure. It uses that capability to define VRF and
VLAN data types in modules/vrf.yml and modules/vlan.yml.