Do not set igp.passive or bgp.advertise on links with role = stub
Routing protocol configuration modules used the 'stub' link role to set the
'passive' interface state. The 'stub' link role is also used to trigger the
calculation of a default gateway, which is then used to create static routes on
Linux VMs and containers.

With the introduction of control-plane daemons running on hosts, we can no
longer use the 'stub' link role as a reliable indicator of IGP 'passive'
interface state, as one of the attached hosts could be a routing protocol
daemon.

The calculation of IGP passive state and BGP prefix advertisement state for
links without an explicit _igp_.passive or bgp.advertise parameter is thus
modified as follows:

* Set _igp_.passive on links with role 'passive' (or any other role matching the
  system default)
* Set _igp_.passive on links that have a single router or routing daemon
  _running the specified IGP_ attached

Setting 'bgp.advertise' parameter is even stricter: the 'bgp.advertise'
parameter is not set for any link with more than one router or daemon attached.