Implement topology search paths and use them when deploying custom configs
The search paths used in various parts of the config deployment process
can be specified in defaults.paths directory. These paths are changed
into absolute paths and stored in the 'all' group to be available to
every managed device.

The custom configuration deployment module uses the 'paths.custom'
and 'paths.deploy' variables to make the Ansible playbook more readable
and to fix #1021.