Another Cumulus MTU fix
'ifreload -a' changes interface MTU when specified within 'inet dhcp6'
but ONLY IN A CUMULUS VM. A container built with the exact same version
of ifupdown2 requires MTU on the raw interface definition (and that
prevents dhcpv6 client to start in a VM).

Workaround: specify MTU on raw interface for containers but not for VMs.

I'm sick and tired of this ****