Bug fix: touch lock file in 'netlab initial' to extend 'netlab validate' delays
The netlab lock file is used to measure the wait times in 'netlab validate'.
That file is touched at the end of the 'netlab up' process, which is usually
OK as 'netlab up' calls 'netlab initial' before that.

However, if you execute 'netlab up --no-config' followed by 'netlab initial',
the lock file does not reflect the time at which the lab was configured,
and the wait times in 'netlab validate' might not be long enough.

Changing the timestamp of the lock file at the end of 'netlab initial'
process fixes that challenge.