Bug fix: stop 'netlab down' complaints with double '--force' option
'netlab down' complains if you try to use it to do a cleanup of a
directory that has lab instance ID identical to another running lab.

The complaints are usually welcome, but not in the CI/CD environment.
This fix changes the '--force' option into a counter, allowing you
to apply even more force to the shutdown process. With enough force
(counter == 2), the complaints stop.

The 'device-module-test' uses this functionality to prevent infinite
blocks if it's started while another lab is running.