Add extra wait time for Junos vPTX in 'netlab validate'
Some devices can be configured before their interfaces are ready
(they just remember the configuration and eventually apply it).
It would be a shame to wait for them during 'netlab initial' as
the rest of the lab can be configured, but they throw a wrench
in the validation process because we have to extend all the timeouts
just to cope with them.

QDS workaround: add extra validation delay configurable in the
features.initial.delay for slow-to-become-sober devices. That extra
delay is used to extend the wait time of the first test that has
'wait' parameter.