Check device-specific Ansible collections during 'netlab up'
The 'devices' module contains code that fetches the installed
Ansible collections and allows device-specific quirks code to
check for their presence. The 'process_config_sw_check' is
executed early in 'netlab up' processing unless the '--no-config'
flag is set, and in 'netlab initial' unless it's invoked with
the '--output' option.