Implement 'valid_with' data type validation
Some attributes are only valid with a subset of other attributes.
For example, it makes no sense to specify a prefix belonging to
a pool plus an IPv4 subnet.

This commit adds '_valid_with' data type checker: a list of other
attributes this one is valid to be used with. If the list is empty,
then you can obviously use the attribute only by itself.