Remove 'netlab_intf_hosts' variable; host names must include interface names
If we define multiple IP address for a hostname in /etc/hosts, Linux picks
a random IP address (not necessarily the first one) when doing ping or
traceroute. To ensure the CLI commands always use the loopback IP address,
we have to include interface names in IP addresses.

Also, for hosts, the hostname has to be before the interface/hostname to
get the correct name during the reverse IP lookup.