Unify /etc/hosts file handling across CL,FRR and Linux VMs and containers (#1317)
The hosts file handling is done exclusively in the 'linux/hosts.j2' template
with the contents being generated in the 'linux/hosts-common.j2' template.

These templates are used in CL, FRR, and Linux initial configuration templates
as well as (using symlinks) in the hosts.j2 templates used to generate clab_files
that are mapped into the container /etc/hosts files.

The content-generation template has been rewritten from scratch (the previous
one generated invalid file format), and deals correctly with multiple IP addresses
per host. 

If we define multiple IP address for a hostname in /etc/hosts, Linux picks
a random IP address (not necessarily the first one) when doing ping or
traceroute. To ensure the CLI commands always use the loopback IP address,
we have to include interface names in IP addresses.

Also, for hosts, the hostname has to be before the interface/hostname to
get the correct name during the reverse IP lookup.