Use 'netlab_ssh_args' for additional SSH arguments
Getting SSH to work with Cisco IOS is a nightmare extraordinaire.
Instead of fighting with ~/.ssh/config, the 'netlab_ssh_args'
variable contains the necessary parameters. They are added to the
'ssh' command in readyness check task list and in the
'netlab connect' command.

Ansible does not need a workaround, the ansible-pylibssh package
works like charm.