Fix VLAN attached interfaces that have gateway: True
For routed interfaces, the VLAN module copies some VLAN data into the
interface data for when preparing VLAN access links. That might
include the 'gateway: true' attribute set on a VLAN.

Usually that wouldn't be a problem due to the usual module attribute
propagation rules, but the gateway module is different. The
'gateway: true' is never replaced with the link data and when the
next consumer (link augmentation) tries to use the interface 'gateway'
dictionary it crashes.

This fix adds another step to the initial processing of link/intf
gateway data: copy link data to interface data if intf gateway is
set to True.

It also adds a safeguard: if the link gateway is set to False, the
gateway attribute is removed from all interfaces.