Fix: Unify generation of unique ifindex/ifname values
Netlab guessed the next ifindex value based on the number of interfaces
of specified type. This fix introduces a deterministic generation of
unique ifindex values for all interface types, ensuring that the
interfaces will have unique ifindex values. The ifname uniqueness
is guaranteed only if the user does not specify ifnames.