Fix: accept prefix.allocation link parameter
Contrary to what the documentation implied, you could not set the
prefix.allocation parameter on a link without specifying static
prefixes.

This change parses the link.prefix data, but checks whether it contains
static IP addressing (including empty and False values). If it does not,
the link.prefix data is merged with the pool-allocated prefix.

Fixes #1458