Use bridge domains for VLANs on IOL
VLAN configuration on IOL requires the 'bridge domain' paradigm similar
to CSR, not the old 'bridge-group' one.

With that, all VLAN tests pass apart from '63-vlan-mixed-native' which
does not work on l3-switch (and we cannot use 'router' model on IOL
because it does not have per-VLAN subinterfaces when bridging).