Revert OS10 template to enable VRRPv3 by default, globally
This commit also fixes a corner case in the VRRP template logic:
```
---
module: [gateway]
defaults.device: dellos10
defaults.provider: clab

nodes: [n1,n2]

links:
- n1:
  n2:
  gateway.protocol: anycast # Quirk: only allowed on SVI interfaces, separate issue
  type: lan
- n1:
  n2:
  gateway.protocol: vrrp
  type: lan
```
loop.first won't work in this case, and vrrp config is skipped. Since the gateway module does a good job cleaning up and only populates the gateway.vrrp property globally when vrrp is used on some interface, it is better and simpler to base the logic on that.

Similarly, the hwaddr filter offers an easier way to format the MAC address