Fix: FRR VRF template is invoked without 'vrfs'
FRR is one of the few devices for which someone decided it makes sense
to retain the 'vrf' module even if the device has no active vrfs.

Anyway, whoever did that forgot to test the 'no vrfs' case, and the
configuration template crashed when a node had no active VRFs.

This commit adds a "let's pretend we have VRFs" logic to the FRR
VRF configuration template