Fix: check the libvirt uplink interface name
If the libvirt.uplink link parameter does not specify a valid interface
name (or if it's missing and the server does not have an eth0
interface), Vagrant throws an error during 'vagrant up' processing.

This fix checks the uplink interface name before Vagrantfile is created.
It also introduces several 'exit_on_error' barriers in the output file
creation process to prevent pollution of output directory in case of
errors.

Inspired by #1551