Check source interface in IBGP integration tests
The IBGP probe in the 02-ibgp-ebgp-session BGP integration test is set
to 'passive BGP' on IBGP neighbors, forcing the device under test to
establish the BGP session. Any problem with the source interface will
thus be uncovered.

The same mechanism is applied to 03-ibgp-rr to catch the scenarios
where the device under test would set the source interface only when
being a route reflector.

Fixes #1599