Fix: Document CL 4.X Ansible caveats, use Linux in tests
New versions of Ansible no longer work with Python release 3.7 running
on the managed nodes, affecting Cumulus Linux 4.x. The final fix has
already been implemented, this is a hotfix for release 1.9.2-post1 as
we're hitting this bug in the wild. It documents the caveat and uses
Linux as the device in the Libvirt and Virtualbox initial tests.