Generic host static routing framework
This commit prepares the static routes that have to be configured
on lab devices not running a routing protocol (hosts). It mirrors
the functionality currently implemented in the Linux configuration
template but makes it generic and ensures the static routes always
have a next hop if at least one interface has a default gateway.

The data structures for all hosts in a lab topology are significantly
modified:

* The 'gateway' data structure is always copied from the link to the
  hosts to ensure the next step has all the relevant information
  available
* The host data contains 'routing.static' list of static routes
  pointing to the first usable gateway.ipv4 address.

The change prepares the data structures needed to fix #1641 but
provides a more generic framework that will be used to implement
'network devices as hosts'