Fix: disable 'ambiguous references' Sphinx warning
Some documentation pages use a reference target identical to the
file name. When that target is used within another file in the
same directory, myst-parser gets confused.

This change disables the "I'm confused" warning, assuming that
the only source of the confusion is the overlap between target
names and file names. If that turns out to be incorrect, we will
have to re-enable the warning and rename all offending targets.