Use timeout to kill vrnetlab readiness check SSH sessions
The SSH client used to check vrnetlab readiness can be stuck for
an awfully long when the VM is in just the right phase of getting
ready. This change uses the 'timeout' command (when available) to
kill the check after 5 seconds, allowing the Ansible retry mechanism
to work as expected.