Improve 'unmanaged' node support (#1776)
The unmanaged nodes were designed to be used in rare cases where
a netlab lab connects to external devices. The initial scenarios
assumed libvirt- or clab uplinks on the connecting links, but that
approach broke when creative engineers started using 'unmanaged'
nodes to reduce workload of exceedingly large topologies.

This commit fixes a number of problems that creativity introduced:

* When the 'unmanaged' nodes are used within the lab (as opposed
  to on uplinks), a number of provider configuration mechanisms
  break. The unmanaged nodes thus have to be removed from link
  interfaces, and some 'p2p' links have to be converted to stub
  links (nodes.py)
* The 'p2p' links converted to stub links lack 'bridge' attribute
  on node interfaces. That attribute is copied (again) to interfaces
  during 'what type of link are we dealing with' libvirt processing
  (libvirt.py)

Furthermore, a number of netlab commands (up, down) broke when dealing
with topology snapshot that includes ghost (unmanaged) nodes. These
commands have to redo the ghostbusting process to get to a topology
that is safe to handle.

Similarly, executing 'netlab connect' or 'netlab capture' on an
unmanaged node makes no sense.

Finally, 'netlab status' command produced bogus information for
unmanaged nodes that (by definition) do not have a corresponding
VM or container.