Resolve static route next hops to interfaces (fixes #1822)
Platforms that do not support indirect next hops for static routes
(for example, Linux) might rely on interface field in the nexthop
data to check that the outgoing interface is known.

The interface field was set for next hops toward a directly-connected
node, but not for directly-connected IP addresses. This fix adds
a function that resolves interfaces for directly-connected IP next hops