Implement 'normalize' configuration step (fixes #1825)
The 'normalize' configuration step is executed before the 'initial'
step. It's triggered with the 'initial' tag (so it's always executed
together with 'initial' step) and uses an extra variable to signal
the deploy-module task list that it's OK if the template is missing
(as it would be for most devices).

The commit also includes the normalization templates for EOS and IOS
layer-2 images