Make LDP data structures similar to other IGPs (fixes #1855)
Every IGP has 'af' dictionary in node data to indicate which AFs are
enabled for that IGP. There was no 'ldp.af' data, resulting in bogus
'we are removing LDP' messages.

This fix also introduces an optional 'remove_module' parameter of the
'remove_unused_igp' function because we cannot remove the 'ldp' module
(it's a sub-function or MPLS module but pretends to be an IGP).