Use Vagrant private key instead of username for FRR (#1963)
This is a proof-of-concept for the 'use Vagrant private key' VM
authentication. FRR is used as an example because it already uses
Debian Bookworm; other devices that could use the same approach are
Cumulus (4/5), VyOS, and Linux.

We're adding private-key-based-authentication functionality to netlab
while retaining the password-based authentication that might be
needed for network management tools.

The Vagrant provisioning script adds the commands to set the password
for the 'vagrant' user (unless it's not defined) and change the
Debian default SSH authentication policy.

The Debian-specific provisioning steps are stored in a separate
template as we'll need them when migrating Linux node to Debian.