vJunos-switch: ebgp evpn overlay
Found workarounds for eBGP-based evpn overlay. next-hop-unchanged is not applicable per AF, but per neighbor (and "hidden" under the multihop hierarchy).
Workaround for that is to set the nh unchanged for the peer, and then on the ebgp export filter force next hop to self for bgp inet and inet6 routes.

Now passing tests:
* evpn/11-vxlan-ebgp.yml
* evpn/14-vxlan-ebgp-ebgp.yml

Test evpn/13-vxlan-ebgp-allowas.yml is still under investigation.

To be merged after #2002