Treat custom config requests with .j2 suffix differently
The file search path for custom config templates included the bare
filename because the config parameter could include a .j2 suffix.

That can result in a matching directory where we really want to
find the .j2 template within the directory (#2051 provides an
excellent example).

This modification removes the bare config filename from the search
path and modifies the search logic -- if the config filename includes
the '.j2' suffix, then we're searching only for that filename,
otherwise we're searching for all potential combinations of things.

Fixes #2015