Fix the SR Linux BGP AF activation (#2074)
This is a minimal set of changes that ensures the BGP address families
are activated as expected -- the address families that are not
explicitely enabled are disabled.

This approach results in a situation where a BGP neighbor might not have
any address families. That's rejected by SR Linux, so we need an extra
device quirk that cleans up the neighbor transport IP addresses based
on the AFs we'll need with the neighbor.

Fixes #2059