Check source VRF of IPv4 addresses used on unnumbered VRF interfaces
While it's possible to run BGP over unnumbered VRF interfaces borrowing
an IPv4 address over a global loopback (because no IPv4 addresses are
involved), that does not work with most IGPs.

This change (based on work by @jbemmel in #2119/#2120) adds a check
for the source of the unnumbered IPv4 address to VRF-specific IGP
processing.

Replaces #2120, closes #2119