Force DUT to have loopback in BGP community integration test
Running 'bird' in host mode resulted in weird interim flaps that took
a while to stabilize (in the end, everything worked and the test
passed). Changing the DUT mode to 'router' made the test pass. As
we're testing the BGP community propagation, the difference should
not matter (for this test).