Fix double transformation for routing.community objects
The routing.community objects start as a list of rules but get
transformed into a dictionary late in the transformation process to
simplify the config templates.

However, the transformation process was also triggered by a routing
policy reference, and later executed again when we processed the node
clists, resulting in unexpected data structure.

This fix checks whether a clist was already transformed before doing
the transformation. It also fixes the common code used in as-path
and clist import to recognize 'regexp' attribute (which can never
appear in as-path object, so we're OK there ;)

Fixes #2155