Implement static (defaults-driven) module hooks
This PR adds another way of calling module hooks: the "hooks" list in
the module definition. The primary use case are hooks that have to be
called before the topology data is validated and the "modules" list is
finalized.

The "normalize" execute_module_hooks function is called before the
data validation starts, giving the modules using it (planned: routing)
the ability to do pre-validation data transformation.

See the discussion in #2154 for more details (hint: this implements
the 'insane' option described in that discussion)

While implementing that functionality, I also refactored the whole
module hook calling subsystem and removed duplicate code.