Fix: Multi-protocol BFD whacked the IGP AF information (#2210)
The multi-protocol BFD code was referencing igp.af data structures
before they were created, resulting in IGP data with no usable AF.
Had to move the BFD call to a later step in the IS-IS post-transform.

Also, the multi-protocol BFD code created IGP data on interfaces
that were already fully processed, so that bit of code needed
further cleanup.