Fix: 'false' value of an attribute with 'true_value' crashed validation
If we have an attribute with 'true_value' (mostly used in bgp.session),
then the 'false' value indicates the attribute should be removed. That
changes the parent dictionary and the keys() iterator crashed. The
trivial part of this commit uses list(keys()) to avoid that crash.

However, there's a bigger problem if the removed attribute could be
inherited from link- or node data. To cope with that, the validation
code creates _removed_attr list to remember which attributes were
removed due to their 'false' value

Solves the trivial part of #2198