Fix: implement stricter checks on node names on links
The code introduced with netlab components accepted any name in the
link string definition as long as a valid node name was its prefix.
That check was upgraded to 'node name + _ but only for components'

Furthermore, the content of 'links' list is rechecked after the
component expansion to ensure the node names are correct (it's
impossible to do the final check before components are expanded).

Fixes #2273