Fix: SR-OS does not support unnumbered EBGP sessions
One has to admire the incredible audacity of a developer who
sets 'features.bgp.ipv6_lla' flag to True while leaving a
"TODO BGP unnumbered" comment in the configuration template.

Alas, faced with either implementing the missing functionality or
disabling the feature, I decided to take the easy way out.

A brave soul wanting to revert this decision is most welcome to
submit a working PR.