SR-OS: Set correct port mode for all VLAN combinations
A combination of SVI VLAN and routed VLAN on the same interface
requires "hybrid" port mode due to the way routed VLAN interfaces
are configured on SR-OS (they are in "router Base", not in IES
service).

This commit moves the calculation of port mode into SR-OS device
module, making it much easier to understand and debug than the
attempts to reverse-engineer it in the Jinja2 template.