Fix: alt-VTEP test is meaningless if traffic starts from DUT
The '08-vxlan-alt-vtep' test is supposed to check whether the device
under test uses an alternate IP for the VXLAN VTEP. However, if we
start the ping from a host attached to DUT, the other VTEPs learn the
"incorrect" DUT VTEP and everything works.

The VXLAN traffic must originate from the probe node if we want to
test the DUT acceptance of VXLAN traffic on an alternate IP address.