Add end-to-end path check to OSPF 'network type' integration tests
I mistakenly believed that a mismatch in OSPF network type would
result in a failure to establish an OSPF adjacency. That's not the
case; the adjacency is established, but the LSAs are useless for
end-to-end path calculation (there's a missing link between type-2
LSA and one of the routers).

The integration tests were thus adjusted to check the bidirectional
path between X1 and X2.