Fix: Use a workaround to do a real clone of the topology data
Output modules or tool templates could inadvertently modify the
topology data, so we're creating a copy of the topology before
calling them. Unfortunately, the Box .copy() function does not give
us a completely independent copy of the data, forcing us to
use a dirty hack: convert Box to dict and back to Box.

Fixes #2465