Fix: Simplify SR-OS IS-IS template with 'configure_router' macro (#2475)
Most SR-OS configuration templates used 'declare_router' macro
that would select the correct service configuration path based on
interface data, leading to an awkward recreation of interface data
in case we really wanted to configure a router, not an interface.

This fix adds another layer of indirection (proving RFC 1925) in
front of the 'declare_router' macro to make the routing protocol
configuration templates a bit cleaner.