Fix: no underscores in topology.name generated from input directory
Underscores in topology.name are used by clab to generate hostnames,
and some Linux distros don't grasp that a clumsy person might have
underscores in hostnames.

This fix ensures there are no weird characters in topology.name that
is generated from the input directory (and 'default' is used when
nothing is left after the cleanup). We don't check what the user
specifies in the 'name' argument (they should know better ;)

Fixes #2424