Fix: Check whether a tool was successfully started
This fix tracks the status of 'tool.up' command(s) as well as the
output generated by Docker. If the output contains 'warning...platform'
text, we might be dealing with a container started on an incompatible
platform, in which case we wait a bit (because it will crash very soon),
check its status, and report the outcome.

The fix also adds better error reporting on tool start/stop failures
and partial-success reporting.

Closes #2456