Fix: Handle the messed-up status of packages on Ubuntu derivatives (#2507)
Pop!_OS supposedly has the 'linux-modules-extra' package, but rejects
to install it with a weird error messsage that doesn't even produce
a proper exit code. This patch uses a slightly different approach to
figuring out if the package is there to be installed.

Tested on Pop!_OS 22.04, Ubuntu 24.04 and Debian Bookworm