Fix: Activate VRF IGP interface only when there are viable neighbors (#2524)
The original VRF IGP code checked whether the IGP should be active in
the VRF based on whether adjacent nodes were running the same IGP, but
not whether they could form an IGP adjacency with the node.

This fix is not perfect, but better than the original -- when checking
whether a neighbor is a viable IGP neighbor, it checks not just the
IGP module (for example, are both running OSPF), but also whether the
two devices have any AF in common on the shared link. That results in
IGP not being run between a router connected to a switch with L2-only
VLAN (but might still result in IGP being active if the switch has
an IP address but does not run IGP in that VRF).

The relevant transformation test was modified to:

* Run OSPF in the underlay
* Put VLANs into a VRF on the router-on-a-stick (to trigger the
  VRF interface migration code)
* Have a single OSPF router in the overlay (the original test had
  two routers, so we'd always get OSPF adjacencies).

Fixes #2275 (mostly, without going into too many edge cases)