Fix: enable device-specific attributes regardless of lab devices
When we added device-specific attributes (eos.serialnumber), we
enabled them only when the device was used in lab topology, making it
impossible to use the same topology with other devices.

This fix enables device-specific attributes regardless of what devices
are used in the current incarnation of the lab topology, allowing you
to specify device-specific data in generic topologies.