Fix: device-specific attributes like ios.debug can be global
When a device-specific is defined as a global attribute, it has to be
applied to every device in the lab topology (similar to what we're
doing with the propagated module settings).

This fix implements the simple merge-from-global behavior. I had to
change the parameters of the 'augment_node_device_data' function
because it needs access to topology data.