VyOS: Implement configuration templates using only 'set' commands
VyOS configuration templates are vbash scripts that include a lot of
setup and cleanup code. The same code is replicated in every template.
That is not good by itself, but it doesn't allow the users to use
simple custom config templates with just the 'set' commands.

This change implements configuration deployment similar to what
we're doing for FRR: if the configuration code includes 'vbash',
it's executed as-is, otherwise it's wrapped in a standard vbash
script.

Also changed:

* The 'are we ready' check has been moved from the 'deploy config'
  task list to 'readyness check' task list
* The 'fetch config' task list fetches device configuration as
  'set' commands, allowing it to be used directly as a custom config
  template. That should allow 'netlab up --reload' to work with
  VyOS.
* BGP and IS-IS templates have been cleaned up as a proof-of-concept