Set 'rr_client' and 'next_hop_self' attributes on BGP neighbor data (#2575)
The BGP neighbor 'rr_client' and 'next_hop_self' attributes will make
the configuration templates much cleaner as we'll avoid all the crazy
logic that is repeated in every template.

Note that the existing attributes are not changed, so there's no rush
to migrate the templates to the new attributes. That can be done at
any time we feel like cleaning them up.

Also: refactor and cleanup the IBGP session generation logic to avoid
duplicate code