Check 'next-hop-self force' on all IBGP sessions in localas-ibgp test
A router a local-as IBGP session MUST force next-hop changes on all
IBGP sessions. This change extends the 08-ibgp-localas test to check
for the proper propagation of IBGP routes between two non-trivial
autonomous systems.

Note: all implementations are failing the extended test at the moment.
We have to fix the neighbor.next_hop_self attribute in the BGP module.