A router with localas_ibgp session MUST change NH on all IBGP routes (#2591)
The routes received over localas_ibgp session have to be reflected to
other IBGP neighbors. The next hop on those reflected routes MUST be
changed because we're not advertising the inter-AS subnet into IGP.

That makes the router with localas_ibgp session unsuitable to be a RR.

This change adds an extra step after the IBGP and EBGP sessions have
been built. It checks for the presence of localas_ibgp session and
sets neighbor.next_hop_self to 'all' on all IBGP sessions if needed.