Refactor 'interface_feature_check' in link transformation code
I thought I would need to check more features in that function, and
it was already bloated with all the IPv4 and IPv6 checks, so I
refactored it into 'ipv4 checks', 'ipv6 checks' and 'common code'.

Not only does it look nicer (and easier to read), it's also ready for
further checks. Maybe I'll eventually move the 'is RA implemented'
checks there.