Add missing 'bgp.as_list.name' attribute we lost in the mists of time (#2680)
In a version far far ago, the bgp.as_list entry could have a 'name'
attribute that was used in topology and BGP graphs. That attribute was
lost once we implemented the strict data checking (because nobody
remembered it at the time), and the graph refactoring done in 25.09
borked its use in graphs.

This PR is fixing both blunders and closes #2675