Remove intra-VRF EBGP sessions with other PE-devices
An intra-VRF EBGP session should not be established between
PE-device running EVPN-over-EBGP when they're in the same EVPN
instance.

This commit adds a 'remove VRF EBGP PE-device neighbors' function
to the EVPN cleanup hook. That function performs a number of tests
before deciding to remove an intra-VRF EBGP session:

* Both devices must run EBGP session in the same VRF. This preserves
  loopback EBGP sessions between VRFs
* Only EBGP sessions between EVPN-enabled devices are removed
* The two EVPN-enabled devices must be in the same evpn.domain -- this
  check enables multi-pod/multi-site designs
* The EBGP session must run over an EVPN-controlled SVI interface --
  this check preserves the in-VRF EBGP sessions in Inter-AS Option A
  designs

Fixes #2477