Simplify reporting/graphing/inspecting workflow
The 'graph', 'report' and 'inspect' commands were taking their data
from the pickled snapshot, requiring you to start a lab or doing
'netlab create' just to generate a graph or a report.

This change makes that workflow more streamlined: the input data is
loaded from the pickle file (if exists), otherwise the 'topology.yml'
file is transformed and used.

The CLI flags allow you to specify a lab instance to get the pickle
file from, the name of the pickle file, or the name of the lab topology